+++ MLEM/Font/GenericSpriteFont.cs
@@ -1,5 +1,4 @@
 using System.Linq;
-using System.Text;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using MLEM.Extensions;
@@ -37,14 +36,9 @@ namespace MLEM.Font {
             return this.Font.MeasureString(c.ToCachedString()).X;
         }
 
-        /// <inheritdoc/>
-        public override void DrawString(SpriteBatch batch, string text, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
-            batch.DrawString(this.Font, text, position, color, rotation, origin, scale, effects, layerDepth);
-        }
-
-        /// <inheritdoc/>
-        public override void DrawString(SpriteBatch batch, StringBuilder text, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
-            batch.DrawString(this.Font, text, position, color, rotation, origin, scale, effects, layerDepth);
+        /// <inheritdoc />
+        protected override void DrawChar(SpriteBatch batch, char c, string cString, Vector2 position, Color color, float rotation, Vector2 origin, Vector2 scale, SpriteEffects effects, float layerDepth) {
+            batch.DrawString(this.Font, cString, position, color, rotation, origin, scale, effects, layerDepth);
         }
 
         private static SpriteFont SetDefaults(SpriteFont font) {